omrelp: Add timeout, so shutdown does not get stuck
By default, omrelp is popular for being extremely reliable
with no message loss in any cases. By default, the action
will try for 90 seconds to deliver the content to the server.
However, if it is during shutdown, then it be running for 90 seconds
after initiating a shutdown event. But this is beyong the threshold
system handles by default and it will think that rsyslog is not able
to properly shut down, thus rsyslog will be marked as a failed service.